Move subcircuit propagate code inside Behaviour implementation
