+++ src/core/simulation/circuit.rs
@@ -1,7 +1,6 @@
 use serde::{Deserialize, Serialize};
 
 use crate::core::simulation::component::{Component, ComponentIdx, ComponentModel, Tick};
-use crate::core::simulation::components::subcircuit::Subcircuit;
 use crate::core::simulation::pin::{Direction, PinIdx};
 use crate::core::simulation::value::operations::assign;
 use crate::core::simulation::wire::{Wire, WireIdx};
@@ -71,31 +70,7 @@ impl Circuit {
             }
 
             for component in first.iter() {
-                if let ComponentModel::Subcircuit(Subcircuit::Instantiated(circuit)) = &component.component {
-                    let mut initial_components = vec![];
-
-                    for (component_pin, circuit_pin) in circuit.input_pins.iter() {
-                        let pin_comp = circuit.components.get(*circuit_pin).unwrap();
-                        if let ComponentModel::InputPin(p) = &pin_comp.component {
-                            initial_components.push(pin_comp);
-
-                            p.value.set(
-                                component.pins.get_value(*component_pin)
-                            );
-                        }
-                    }
-
-                    circuit.propagate(initial_components);
-
-                    for (component_pin, circuit_pin) in circuit.output_pins.iter() {
-                        let pin_comp = circuit.components.get(*circuit_pin).unwrap();
-                        if let ComponentModel::OutputPin(p) = &pin_comp.component {
-                            component.pins.set_value(*component_pin, p.value.get());
-                        }
-                    }
-                } else {
-                    component.propagate();
-                }
+                component.propagate();
             }
 
             // println!("{:?}", first);
