+++ src/core/simulation/components/subcircuit/mod.rs
@@ -4,7 +4,7 @@ use std::rc::Rc;
 use serde::{Deserialize, Serialize};
 
 use crate::core::simulation::circuit::Circuit;
-use crate::core::simulation::component::{Component, ComponentModel, ComponentPins, ComponentProperties};
+use crate::core::simulation::component::{Behaviour, Component, ComponentModel, ComponentPins, ComponentProperties};
 use crate::core::simulation::components::subcircuit::Subcircuit::NotInstantiated;
 use crate::core::simulation::pin::Pin;
 use crate::serde::project::SavedCircuit;
@@ -16,6 +16,36 @@ pub mod output_pin;
 pub enum Subcircuit {
     Instantiated(Rc<Circuit>),
     NotInstantiated(String),
+}
+
+impl Behaviour for Subcircuit {
+    fn propagate(&self, pins: &ComponentPins, _properties: &ComponentProperties) {
+        debug_assert!(matches!(self, Subcircuit::Instantiated(_)));
+
+        if let Subcircuit::Instantiated(circuit) = self {
+            let mut initial_components = vec![];
+
+            for (component_pin, circuit_pin) in circuit.input_pins.iter() {
+                let pin_comp = circuit.components.get(*circuit_pin).unwrap();
+                if let ComponentModel::InputPin(p) = &pin_comp.component {
+                    initial_components.push(pin_comp);
+
+                    p.value.set(
+                        pins.get_value(*component_pin)
+                    );
+                }
+            }
+
+            circuit.propagate(initial_components);
+
+            for (component_pin, circuit_pin) in circuit.output_pins.iter() {
+                let pin_comp = circuit.components.get(*circuit_pin).unwrap();
+                if let ComponentModel::OutputPin(p) = &pin_comp.component {
+                    pins.set_value(*component_pin, p.value.get());
+                }
+            }
+        }
+    }
 }
 
 impl Subcircuit {
